Isolate Feishu integration tests behind their own daemon shard
Feishu doctor and CLI coverage had grown large enough that bundling them into
`daemon_channels` kept unrelated channel-surface work paying for Feishu-only
compilation and fixture setup. Split those tests into a dedicated `daemon_feishu`
shard and move the duplicated grant/config setup into one narrow shared fixture
module so the channel shard can stay focused on catalog, bridge, and serve
surfaces.

Constraint: Shard routing must stay derivable from the daemon test manifest and module layout
Rejected: Keep Feishu under daemon_channels | unrelated channel changes still compile and run Feishu coverage
Confidence: high
Scope-risk: moderate
Directive: Keep CLI-only HTTP mocks and account/base-url variants local to feishu_cli; only share canonical Feishu config and grant fixtures
Tested: ./scripts/cargo-local-toolchain.sh test --locked -p loong --features test-support --test daemon_feishu
Tested: ./scripts/cargo-local-toolchain.sh test --locked -p loong --features test-support --test daemon_channels
Tested: ./scripts/cargo-local-toolchain.sh test --locked -p loong --test integration
Not-tested: Remote GitHub Actions execution
