+++ crates/daemon/tests/integration/mod.rs
@@ -15,6 +15,7 @@ mod chat_cli;
 mod cli_tests;
 mod doctor_feishu;
 mod feishu_cli;
+mod feishu_test_fixtures;
 mod gateway_api_acp;
 mod gateway_api_events;
 mod gateway_api_health;
@@ -48,6 +49,7 @@ mod spec_runtime;
 mod spec_runtime_bridge;
 mod status_cli;
 mod tasks_cli;
+pub(crate) use feishu_test_fixtures::*;
 pub(crate) use managed_bridge_fixtures::*;
 mod trajectory_export_cli;
 mod validate_config_root;
